requirements.txt
httpx==0.26.0
pdfplumber==0.11.9
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.20
python-dotenv==1.1.1
